refactor: rename key_verification_flow to key_verification_controller
The file now holds KeyVerificationController, not a flow mixin, so the
name should match.

Refs #623
