+++ lib/features/e2ee/widgets/key_verification_content.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
-import 'package:kohera/features/e2ee/widgets/key_verification_flow.dart';
+import 'package:kohera/features/e2ee/widgets/key_verification_controller.dart';
 import 'package:kohera/features/e2ee/widgets/qr_verification_views.dart';
 import 'package:kohera/shared/widgets/kohera_loader.dart';
 import 'package:matrix/encryption.dart';
